Rearrange the files and collapse .h files in SimCalorimetry/HcalTrigPrimProducers
